Properly package-initialize
Packages such as use-package can't really deal with
installed-but-not-activated packages properly. We need to activate
everything so that they are on the load path — actually loading them can
still be deferred.

What this was originally trying to solve was not enabling packages that
aren't listed in package-list, for which there are better solutions.