Make byte-compiling init.el work correctly
The tricky thing here is that we want to eval-and-compile the
use-package initialization because use-package works at compile time, so
we need to package-install it at compile time if it's not yet there. But
since it works /only/ at compile time, at runtime we would see that
my-used-packages is empty and thus consider everything an orphan. So
make a little magic-looking macro that basically emits the code to
populate my-used-packages at runtime based on its compile time contents.