Remove package: smooth-scrolling
It breaks at least scrolling backwards beyond the window edge: e.g. 100k
might move only one line instead of 100.

This was also annoying to debug because the package autoloaded and
applied its advice even without the use-package, just as long as it was
installed. This seems like nasty misbehaviour to me.