eyebrowse: update key bindings
I noticed that orgtbl-mode, which e.g. feature-mode enables, overrides
eyebrowse's default keymap prefix (C-c C-w) and breaks all kinds of
things as a result. The M-number bindings and C-" should be safer in
that regard.

And use eyebrowse-setup-evil-keys instead of our own thing, it also
binds a few other ones that could be useful.